+++ rotograd/__init__.py
@@ -1,5 +1,5 @@
-from .rotograd import VanillaMTL, RotoGrad, RotoGradNorm, cached
+from .rotograd import VanillaMTL, RotoGrad, RotoGradNorm, cached, RotateOnly
 
-__version__ = '0.1.4'
+__version__ = '0.1.5'
 
-__all__ = ['VanillaMTL', 'RotoGrad', 'RotoGradNorm', 'cached']
+__all__ = ['VanillaMTL', 'RotateOnly', 'RotoGrad', 'RotoGradNorm', 'cached']